:recycle: move more internals into v1/v2
