+++ src/V1/Client.php
@@ -16,14 +16,14 @@ use Mindee\CustomSleepMixin;
 use Mindee\Error\ErrorCode;
 use Mindee\Error\MindeeApiException;
 use Mindee\Error\MindeeException;
-use Mindee\Error\V1\MindeeV1ClientException;
-use Mindee\Error\V1\MindeeV1HttpException;
 use Mindee\Input\InputSource;
 use Mindee\Input\LocalInputSource;
 use Mindee\Input\LocalResponse;
 use Mindee\Input\PageOptions;
 use Mindee\V1\ClientOptions\PredictMethodOptions;
 use Mindee\V1\ClientOptions\WorkflowOptions;
+use Mindee\V1\Error\MindeeV1ClientException;
+use Mindee\V1\Error\MindeeV1HttpException;
 use Mindee\V1\Http\Endpoint;
 use Mindee\V1\Http\MindeeApi;
 use Mindee\V1\Http\MindeeWorkflowApi;
