+++ tests/ClientTest.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 use Mindee\ClientOptions\PollingOptions;
 use Mindee\Error\MindeeApiException;
 use Mindee\Error\MindeeMimeTypeException;
-use Mindee\Error\V1\MindeeV1HttpException;
 use Mindee\Input\LocalResponse;
 use Mindee\Input\PageOptions;
 use Mindee\Input\PathInput;
 use Mindee\Input\UrlInputSource;
 use Mindee\V1\Client;
 use Mindee\V1\ClientOptions\PredictMethodOptions;
+use Mindee\V1\Error\MindeeV1HttpException;
 use Mindee\V1\Product\Generated\GeneratedV1;
 use Mindee\V1\Product\Invoice\InvoiceV4;
 use Mindee\V1\Product\InvoiceSplitter\InvoiceSplitterV1;
